+++ mbsimgui/mbsimgui/signal_property_dialog.h
@@ -143,7 +143,7 @@ namespace MBSimGUI {
       xercesc::DOMElement* initializeUsingXML(xercesc::DOMElement *parent) override;
       xercesc::DOMElement* writeXMLFile(xercesc::DOMNode *element, xercesc::DOMNode *ref=nullptr) override;
     protected:
-      ExtWidget *function;
+      ExtWidget *function, *ifv, *ifd, *isd;
   };
 
   class ContactSensorPropertyDialog : public SensorPropertyDialog {